Move ParsedMessage into the clp namespace.
