+++ components/core/src/clp/ParsedMessage.hpp
@@ -1,10 +1,11 @@
-#ifndef PARSEDMESSAGE_HPP
-#define PARSEDMESSAGE_HPP
+#ifndef CLP_PARSEDMESSAGE_HPP
+#define CLP_PARSEDMESSAGE_HPP
 
 #include <string>
 
-#include "TimestampPattern.hpp"
+#include "../TimestampPattern.hpp"
 
+namespace clp {
 /**
  * ParsedMessage represents a (potentially multiline) log message parsed into 3 primary fields:
  * timestamp, timestamp pattern, and content.
@@ -68,5 +69,6 @@ private:
     size_t m_orig_num_bytes;
     bool m_is_set;
 };
+}  // namespace clp
 
-#endif  // PARSEDMESSAGE_HPP
+#endif  // CLP_PARSEDMESSAGE_HPP
