+++ components/core/src/clp/Utils.hpp
@@ -12,7 +12,7 @@
 #include "../Defs.h"
 #include "../ErrorCode.hpp"
 #include "../FileReader.hpp"
-#include "../ParsedMessage.hpp"
+#include "ParsedMessage.hpp"
 
 namespace clp {
 /**
